even more of week 4

+++ 3067/increase.py
@@ -1,8 +1,8 @@
 """increase"""
 
-num1 = int(input())
-num2 = int(input())
-num3 = int(input())
+num1 = float(input())
+num2 = float(input())
+num3 = float(input())
 
 if num2 - num1 > 0 and num3 - num2 > 0:
     print("increasing")
